First draft of Harvest, aka Facai v2
On the train to Berlin, on the way to BobKonf and Racketfest, 2022-03-13

Starting from scratch, but with some bits copied over from Facai, like the path
matching. This rewrite achieves a number of things:

Most noticable is that factories now take an argument vector, in other words,
they act more like functions. They are still maps/records, but the template key
is now always a function. This means that within a factory it's all "just
Clojure", you can call other functions and do whatever you like as you normally
would.

Harvest will evaluate this function to get the actual template back, a value,
most commonly a map. It will then as before recurse over this structure, further
expanding/building factories as it encounters them. It will not, however, call
functions for you. In the end this means a lot less wrapping stuff in lambdas.

When calling a factory as a function, this now returns the factory, but
including the captured argument vector, for application during the build
process. So we have essentially the "deferring" as in Facai, but without it
having to be a special case.

The build process is simplified by introducing two dynamic variables,
`*context*` and `*buildset*`. These replace the context map which was passed all
the way through in Facai. While it's compelling to have the process build up
this context map in a functional way, but there is a tension between passing the
context "forward" (so results can be accumulated), and having it be stack based,
passed downwards but not upwards. This new implementation separates those two
concerns. Things like rules, which are passed from higher factories down to
their children (but not vice versa), go into *context*, the resulting values are
added into `*buildset*`, which is mutable (a volatile that lives for the
duration of the build process). This also removes the distinction between
`value` and `linked`. What was previously `value` is now simply the root in the
buildset.

Calling factories does nothing special, we simply pass the args
through (factories are functions now, remember?). To do things like value
overrides (with) or declarative overrides of nested values (rules), there are
helpers which assoc these things onto the factory, and they are then honored
during the build process.

All in all this rewrite gets rid of a few arbitrary distinctions:
- deferred vs actual factory
- args passed to factory vs args passed to build

while separating concepts which previously were conflating:
- passing context down vs building up values for return